UI Components refactoring
