+++ BitStream/Views/Components/CustomProgressView.swift
