+++ BitStream/styles/GradientButtonStyle.swift
@@ -1,0 +1,40 @@
+//
+//  GradientButtonStyle.swift
+//  BitStream
+//
+//  Created by GICHUKI on 05/10/2025.
+//
+
+import SwiftUI
+
+struct GradientButtonStyle: ButtonStyle {
+    let isEnabled: Bool
+    
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .foregroundColor(.white)
+            .background(
+                Group {
+                    if isEnabled {
+                        LinearGradient(
+                            colors: [Color.blue, Color.purple],
+                            startPoint: .leading,
+                            endPoint: .trailing
+                        )
+                        .opacity(configuration.isPressed ? 0.8 : 1.0)
+                    } else {
+                        Color.gray.opacity(0.3)
+                    }
+                }
+            )
+            .cornerRadius(12)
+            .shadow(
+                color: isEnabled ? Color.blue.opacity(0.3) : Color.clear,
+                radius: configuration.isPressed ? 4 : 8,
+                x: 0,
+                y: configuration.isPressed ? 2 : 4
+            )
+            .scaleEffect(configuration.isPressed ? 0.98 : 1.0)
+            .animation(.spring(response: 0.3, dampingFraction: 0.7), value: configuration.isPressed)
+    }
+}
